Conceptual programme change
Transfer functions to their respective classes ('checkClassificationRule' and 'splitStringOfClassificationRules' to ClassificationRules class;
'checkRecords' and 'splitStringOfRecords' to the Records class;
'classificationRecordsByRule' and 'buildStringFromResult' to the Result class;
'writeStringToFile' and 'unpackTextFile' to the created FileHandling class)
Creating the FileHandling class.
Point function edits. Removing or changing variables and class methods.
Adding checks in main for input file emptiness.
Adding checks for existence of path to input and output files.
General formatting of the program code.