Structural changes to function calls
The 'classificationRecordsByRule' function now calls the 'splitStringOfClassificationRules' and 'splitStringOfRecords' functions.
Containers of classes Records, ClassificationRules and Result are now stored in private class Result instead of main.
The 'buildStringFromResult' function is now called in the 'classificationRecordsByRule' function and is the return value of the function.
The parameters passed to the 'classificationRecordsByRule' function have been changed. It now accepts two QString (input text of records and classification rules).
Spot change in the operation of loops and conditions in the main 'classificationRecordsByRule' function