Structural changes to function calls
The 'splitStringOfClassificationRules' function now calls the 'checkClassificationRule' function and returns an error string or text to the main 'classificationRecordByRule' function: "It's OK!". It's the same with the 'splitStringOfRecords' function. Removed from the main function call for checking functions. Added checks on the returned text of the 'splitStringOfClassificationRules' and 'splitStringOfRecords' functions, so that the main function 'classificationRecordByRule' will not continue working after returning an error in the rules and/or records text.